+++ src/module/free_gimbal/mod_free_gimbal.hpp
@@ -44,9 +44,9 @@ class FreeGimbal {
  private:
   float yaw_out_;
 
-  float last_wakeup_ = 0.0f;
+  uint64_t last_wakeup_ = 0;
 
-  float now_ = 0.0f;
+  uint64_t now_ = 0;
 
   float dt_ = 0.0f;
