+++ src/module/gimbal/mod_gimbal.hpp
@@ -76,9 +76,9 @@ class Gimbal {
   static void DrawUIDynamic(Gimbal *gimbal);
 
  private:
-  float last_wakeup_ = 0.0f;
+  uint64_t last_wakeup_ = 0;
 
-  float now_ = 0.0f;
+  uint64_t now_ = 0;
 
   float dt_ = 0.0f;
