+++ src/module/ore_collect/mod_ore_collect.hpp
@@ -68,9 +68,9 @@ class OreCollect {
 
   float dt_ = 0.0f;
 
-  float last_wakeup_ = 0.0f;
+  uint64_t last_wakeup_ = 0;
 
-  float now_ = 0.0f;
+  uint64_t now_ = 0;
 
   Mode mode_ = RELAX;
